pyproject.toml

[project]
name = "glue-inspector"
version = "0.1.0"
description = "AWS Glue Job Inspector - Analyze dependencies and configurations"
authors = [
    { name = "Jacob Verhoeks", email = "jacob.verhoeks@gmail.com" }
]
requires-python = ">=3.9"
readme = "README.md"
license = { file = "LICENSE" }

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "boto3>=1.26.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "cyclonedx-python-lib>=3.1.5",
    "packaging>=23.0",
    "urllib3>=1.26.20",
    "cyclonedx-py>=1.0.1",
    "markdown-table-generator>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0"
]

[project.scripts]
glue-inspector = "glue_inspector.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/glue_inspector"]


[tool.ruff]
lint.select = ["E", "F", "B", "I"]
lint.ignore = []
line-length = 120

[tool.isort]
profile = "black"
line_length =  120
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=glue_inspector"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "bandit>=1.8.0",
    "pytest>=8.3.4",
]

[tool.bandit]
targets = ["src/glue_inspector"]
skips = ["B404","B603","B607"]
